adds methods filter_country, print_country to scraper_logic
creates a new variable with attr_accessor so that the main.rb can acess the methods in the scraper_logic
