+++ lib/scraper_logic.rb
@@ -9,12 +9,65 @@ require 'tty-spinner'
 require 'tty-link'
 require 'byebug'
 
-require_relative '../bin/main'
 
 class Scraper
-
-    def print_all
-        $countries_stats.each do |country|
+  def print_country(country)
+    puts Terminal::Table.new(
+      rows: [
+        [country[:name], 
+        country[:total_cases], 
+        country[:new_cases],
+        country[:total_deaths],
+        country[:new_deaths],
+        country[:total_recovered],
+        country[:new_recovered],
+        country[:active_cases],
+        country[:serious_critical],
+        country[:total_cases_per_1_milion]]
+      ],
+      headings: [
+        'Country'.bold.blue,
+        'Total'.red,
+        'New cases'.bold.blue,
+        'Total deaths'.bold.red,
+        'New deaths'.red,
+        'Total reco'.bold.green,
+        'New reco'.bold.green,
+        'Active cases'.red,
+        'Serious/critical'  .red,
+        'Total cases/1M'.bold.blue,
+        
+      ],
+      style: {
+        border_i: '+'
+      }
+     )
+  
+     puts Terminal::Table.new(
+       rows: [
+        [country[:deaths_per_1_milion],
+        country[:total_tests],
+        country[:tests_per_1_milion],
+        country[:population],
+        country[:continent] ]
+       ],
+       headings: [
+        'Total deaths/1M'.red,
+        'Total tests'.bold.blue,
+        'Total tests/1M'.bold.blue,
+        'Population'.bold.green,
+        'continent'.bold.green
+       ],
+       style: {
+        border_i: '+'
+      }
+     )
+      puts ""
+    end
+  
+    
+    def print_all(countries_stats)
+        countries_stats.each do |country|
           print_country(country)
         end
     end
@@ -25,17 +78,16 @@ class Scraper
        country[0]
     end
 
-  def ask_user
-    
+  def ask_user(countries_stats)
     prompt = TTY::Prompt.new
     puts ""
-    user_input = prompt.select("Choose 'All' or a specific country to know stats     about covid-19?", 'All', $countries_stats.each { |country_name| country_name    [:name] }).capitalize
+    user_input = prompt.select("Choose 'All' or a specific country to know stats     about covid-19?", 'All', countries_stats.each { |country_name| country_name[:name] })
             
     case user_input
       when 'All'
-      print_all
+      print_all(countries_stats)
       else
-      print_country(filter_country($countries_stats, user_input))
+      print_country(filter_country(countries_stats, user_input))
     end
     font = TTY::Font.new(:doom)
     puts font.write("thanks for using").light_blue
